AIML/Basic/Image_Tagging/Files/cifar10_project/cifar10_project/requirements.txt
tensorflow>=2.13.0
flask>=3.0.0
flask-socketio>=5.3.0
pillow>=10.0.0
numpy>=1.24.0
eventlet>=0.35.0
